data-type modifications to correspond data-generation process

+++ model/time_encoding.py
@@ -11,8 +11,8 @@ class TimeEncode(torch.nn.Module):
         self.w = torch.nn.Linear(1, dimension)
 
         self.w.weight = torch.nn.Parameter((torch.from_numpy(1 / 10 ** np.linspace(0, 9, dimension)))
-                                           .reshape(dimension, -1))
-        self.w.bias = torch.nn.Parameter(torch.zeros(dimension))
+                                           .float().reshape(dimension, -1))
+        self.w.bias = torch.nn.Parameter(torch.zeros(dimension).float())
 
     def forward(self, t):
         # t has shape [batch_size, seq_len]
